Clean up tensor_util and torch_util
- Merge torch_util into tensor_util to reduce the number of util files
- Remove unused functions in tensor_util.cpp
- Move TensorHash into hash.h
- Move TensorCompare into lazy_graph_executor.cpp
