+++ lazy_tensor_core/lazy_tensor_core/csrc/ops/as_strided.cpp
@@ -3,7 +3,6 @@
 #include <algorithm>
 
 #include "lazy_tensor_core/csrc/tensor_util.h"
-#include "lazy_tensor_core/csrc/torch_util.h"
 #include "lazy_tensors/computation_client/util.h"
 #include "lazy_tensors/shape_util.h"
