+++ lazy_tensor_core/lazy_tensor_core/csrc/ops/cast.cpp
@@ -2,7 +2,6 @@
 
 #include "lazy_tensor_core/csrc/ops/ltc_ops.h"
 #include "lazy_tensor_core/csrc/tensor_util.h"
-#include "lazy_tensor_core/csrc/torch_util.h"
 
 namespace torch_lazy_tensors {
 namespace ir {
