+++ lazy_tensor_core/test/cpp/test_aten_ltc_ts_tensor.cpp
@@ -7,7 +7,6 @@
 #include "cpp_test_util.h"
 #include "lazy_tensor_core/csrc/aten_ltc_bridge.h"
 #include "lazy_tensor_core/csrc/helpers.h"
-#include "lazy_tensor_core/csrc/torch_util.h"
 #include "lazy_tensor_core/csrc/ts_backend/backend_impl.h"
 #include "lazy_tensors/computation_client/metrics.h"
 #include "torch_ltc_ts_test.h"
