+++ torch/csrc/lazy/core/hash.h
@@ -9,8 +9,8 @@
 #include <string>
 #include <vector>
 
+#include <ATen/Tensor.h>
 #include <c10/core/Scalar.h>
-#include <c10/core/ScalarType.h>
 #include <c10/util/int128.h>
 #include <torch/csrc/WindowsTorchApiMacro.h>
 
@@ -84,7 +84,41 @@ static inline hash_t Hash(const c10::Scalar& value) {
   case c10::ScalarType::Bool:
     return Hash(value.toBool());
   default:
-    TORCH_INTERNAL_ASSERT(false, "Unknown scalar type.");
+    TORCH_INTERNAL_ASSERT(false, "Unknown scalar type.", value.type());
+  }
+}
+
+static inline torch::lazy::hash_t TensorHash(const at::Tensor& tensor) {
+  at::Tensor ctensor = tensor.contiguous();
+  int64_t size = ctensor.numel() * ctensor.element_size();
+  switch (ctensor.scalar_type()) {
+    case at::ScalarType::Bool:
+      return DataHash(ctensor.data_ptr<bool>(), size);
+    case at::ScalarType::Byte:
+      return DataHash(ctensor.data_ptr<uint8_t>(), size);
+    case at::ScalarType::Char:
+      return DataHash(ctensor.data_ptr<int8_t>(), size);
+    case at::ScalarType::Short:
+      return DataHash(ctensor.data_ptr<int16_t>(), size);
+    case at::ScalarType::Int:
+      return DataHash(ctensor.data_ptr<int32_t>(), size);
+    case at::ScalarType::Long:
+      return DataHash(ctensor.data_ptr<int64_t>(), size);
+    case at::ScalarType::Float:
+      return DataHash(ctensor.data_ptr<float>(), size);
+    case at::ScalarType::Double:
+      return DataHash(ctensor.data_ptr<double>(), size);
+    case at::ScalarType::BFloat16:
+      return DataHash(ctensor.data_ptr<at::BFloat16>(), size);
+    case at::ScalarType::Half:
+      return DataHash(ctensor.data_ptr<at::Half>(), size);
+    case at::ScalarType::ComplexFloat:
+      return DataHash(ctensor.data_ptr<c10::complex<float>>(), size);
+    case at::ScalarType::ComplexDouble:
+      return DataHash(ctensor.data_ptr<c10::complex<double>>(), size);
+    default:
+      TORCH_INTERNAL_ASSERT(
+          false, "Unsupported scalar type:", ctensor.scalar_type());
   }
 }
